Split types/index.ts barrel file into domain modules (#1225)
- Create types/content.ts (database entity types)
- Create types/user.ts (user data types)
- Create types/explore.ts (feature/browse types)
- Create types/meta.ts (parsed panel content types)
- Reduce index.ts to 7-line re-export barrel
- All existing imports continue to work unchanged
